pyproject.toml
[tool.poetry]
name = "assistant"
version = "0.1.0"
description = "A Smart Assistant for Ideation"
authors = ["Peter <mail@peterhansen.io>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.2"
tiktoken = "^0.3.3"
aiofiles = "^23.1.0"
loguru = "^0.6.0"
langchain = "^0.0.128"
aiohttp = "^3.8.4"
pyzmq = "^25.0.2"
uri = "^2.0.1"


[tool.poetry.group.meta.dependencies]
jinja2 = "^3.1.2"
networkx = "^3.1"
py-cord = "^2.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
